Now have a shop menu open when you talk to the wholesaler Big'n the flame atronach, and its populating items based on his inventory. Lots of other stuff added like other NPC stuff and reworked all that to be SO's. Now to have the player buy the item and be able to display it at their stall and then have customers buy products based on some maths.

+++ Assets/Scripts/Item/ItemSO.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "NewItemType", menuName = "Scriptable Objects/Item/ItemSO")]
@@ -5,11 +6,10 @@ public class ItemSO : ScriptableObject
 {
     public string id;
     public string itemName;
-    [TextArea]
-    public string description;
+    [TextArea] public string description;
     public Sprite itemSprite;
     public int baseValue;
-    public int weight;
+    public float weight;
     public int stackSize;
     public bool isRawMaterial;
     public bool isCraftedGood;
